requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
dateparser==1.1.1
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
numpy==1.22.3
pandas==1.4.2
python-binance==1.0.16
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.27.1
six==1.16.0
termcolor==1.1.0
tzdata==2022.1
tzlocal==4.2
ujson==5.2.0
urllib3==1.26.9
websocket-client==1.3.2
websockets==10.2
yarl==1.7.2
